requirements.txt
torch
numpy
pandas
Pillow
h5py
argparse
sklearn
matplotlib